requirements.txt
asgiref==3.2.10
autopep8==1.5.7
Django==3.0.8
django-environ==0.7.0
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.0.4
mysqlclient==2.1.0
Pillow==8.4.0
pycodestyle==2.8.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.2
djangorestframework-simplejwt==5.0.0
dj-rest-auth==2.1.11
